Create HlsServer GenServer
