+++ camera_to_hls_nerves/lib/camera_to_hls_nerves/application.ex
@@ -1,0 +1,15 @@
+defmodule CameraToHlsNerves.Application do
+  use Application
+
+  def start(_type, _args) do
+    File.rm_rf!("/data/output")
+    File.mkdir!("/data/output")
+
+    children = [
+      CameraToHlsNerves.HlsServer,
+      CameraToHlsNerves.Pipeline
+    ]
+
+    Supervisor.start_link(children, strategy: :one_for_one)
+  end
+end
