+++ camera_to_hls_nerves/lib/camera_to_hls_nerves/hls_server.ex
@@ -1,0 +1,31 @@
+defmodule CameraToHlsNerves.HlsServer do
+  use GenServer
+
+  def start_link(arg) do
+    GenServer.start_link(__MODULE__, arg, name: :hls_server)
+  end
+
+  @impl true
+  def init(_arg) do
+    {:ok, nil}
+  end
+
+  @impl true
+  def handle_info(:playlist_ready, nil) do
+    File.cd!("/")
+    
+    :inets.start()
+
+    :inets.start(:httpd, httpd_options())
+  end
+
+  defp httpd_options() do
+    [
+      bind_address: ~c"0.0.0.0",
+      port: 8000,
+      document_root: ~c".",
+      server_name: ~c"camera_to_hls_nerves",
+      server_root: ~c"/"
+    ]
+  end
+end
